Change return type of xvsnprintf() and xsnprintf() from int to size_t
The vsnprintf(3) standard library function returns int, but only for
historical reasons. It would make more sense for it to return ssize_t.
With xvsnprintf() the return is never negative, so we can use size_t
instead.

Note: this change isn't for the purpose of increasing the maximum
length. It would be insane to use vsnprintf() with strings anywhere
near INT_MAX bytes. It's more so to allow the compiler to generate
better code and to make the purpose of the return type more obvious.

+++ src/debug.c
@@ -75,12 +75,12 @@ void debug_print(const char *function, const char *fmt, ...)
 
     char buf[4096];
     size_t write_max = ARRAY_COUNT(buf);
-    const int len1 = xsnprintf(buf, write_max, "%s: ", function);
+    const size_t len1 = xsnprintf(buf, write_max, "%s: ", function);
     write_max -= len1;
 
     va_list ap;
     va_start(ap, fmt);
-    const int len2 = xvsnprintf(buf + len1, write_max, fmt, ap);
+    const size_t len2 = xvsnprintf(buf + len1, write_max, fmt, ap);
     va_end(ap);
 
     xwrite(fd, buf, len1 + len2);
